Started Multithreaded version of application.
It will now generate an IDF file and attempt to run the EnergyPlus simulation. Once finished, the necessary data (a few numbers from a CSV file) will be extracted and written to a single file. Finally all unnecessary files will be deleted.

This will take more time than before, but we will no longer need large amounts of disk space.

Choosing to do this to save space. Otherwise, we end up possibly millions of IDF files that must be run sitting on disk.  This way, we end up only with